Refactor get_request_ip into utils
