+++ pccommon/pccommon/constants.py
@@ -2,3 +2,8 @@ DEFAULT_COLLECTION_CONFIG_TABLE_NAME = "collectionconfig"
 DEFAULT_CONTAINER_CONFIG_TABLE_NAME = "containerconfig"
 
 DEFAULT_TABLE_TTL = 600  # 10 minutes
+# Headers containing information about the requester's
+# IP address. Checked in the order listed here.
+X_AZURE_CLIENTIP = "X-Azure-ClientIP"
+X_ORIGINAL_FORWARDED_FOR = "X-Original-Forwarded-For"
+X_FORWARDED_FOR = "X-Forwarded-For"
