+++ pccommon/pccommon/utils.py
@@ -2,6 +2,13 @@ import urllib.parse
 from typing import Any, Callable, Dict, Optional, TypeVar
 
 import orjson
+from fastapi import Request
+
+from pccommon.constants import (
+    X_AZURE_CLIENTIP,
+    X_FORWARDED_FOR,
+    X_ORIGINAL_FORWARDED_FOR,
+)
 
 T = TypeVar("T")
 U = TypeVar("U")
@@ -26,3 +33,16 @@ def map_opt(fn: Callable[[T], U], v: Optional[T]) -> Optional[U]:
 def orjson_dumps(v: Dict[str, Any], *args: Any, default: Any) -> str:
     # orjson.dumps returns bytes, to match standard json.dumps we need to decode
     return orjson.dumps(v, default=default).decode()
+
+
+def get_request_ip(request: Request) -> str:
+    """Gets the IP address of the request."""
+
+    ip_header = (
+        request.headers.get(X_AZURE_CLIENTIP)  # set by Front Door
+        or request.headers.get(X_ORIGINAL_FORWARDED_FOR)
+        or request.headers.get(X_FORWARDED_FOR)
+    )
+
+    # If multiple IPs, take the first one
+    return ip_header.split(",")[0] if ip_header else ""
